Integrate bundling and size check processes
